Switch ConnectionsData.DataSize to long, add covering index
Update ConnectionsData and related code to use long for DataSize, supporting larger payloads. Add IX_ConnectionsData_Session_Direction_Size covering index for efficient bandwidth queries. Optimize BandwidthStats query to leverage the new index and sum DataSize in C# for improved performance.

+++ ReverseProxyServer.Core/EventArgs/ConnectionDataEventArgs.cs
@@ -2,11 +2,11 @@
 
 namespace ReverseProxyServer.Core
 {
-    public class ConnectionDataEventArgs(string sessionId, string communicationDirection, StringBuilder rawData, int dataSize) : EventArgs
+    public class ConnectionDataEventArgs(string sessionId, string communicationDirection, StringBuilder rawData, long dataSize) : EventArgs
     {
         public string SessionId { get; init; } = sessionId;
         public string CommunicationDirection { get; init; } = communicationDirection;
         public StringBuilder RawData { get; init; } = rawData;
-        public int DataSize { get; init; } = dataSize;
+        public long DataSize { get; init; } = dataSize;
     }
 }
