+++ ReverseProxyServer.Data/DTO/ConnectionData.cs
@@ -18,14 +18,14 @@ namespace ReverseProxyServer.Data.DTO
         public string? SessionId { get; set; }
         public string? CommunicationDirection { get; set; }
         public string? Data { get; set; }
-        public int DataSize { get; set; }
+        public long DataSize { get; set; }
 
 
         public ConnectionData()
         {
 
         }
-        public ConnectionData(string sessionId, string communicationDirection, string data, int dataSize)
+        public ConnectionData(string sessionId, string communicationDirection, string data, long dataSize)
         {
             this.SessionId = sessionId;
             this.CommunicationDirection = communicationDirection;
